refactor: split db.ts into domain modules
Splits the monolithic db.ts (670 lines) into domain-specific modules
under src/db/. Each module handles one database domain:

- init.ts: schema, migrations, getDb(), initDatabase()
- chats.ts: chat metadata storage and queries
- messages.ts: message storage and retrieval
- tasks.ts: scheduled task CRUD and run logging
- sessions.ts: Claude session management
- groups.ts: registered group CRUD
- router-state.ts: key-value router state

A barrel re-export in db.ts ensures zero changes needed in existing
importers — all `import { ... } from './db.js'` statements continue
to work unchanged.
