+++ src/db/index.ts
@@ -1,0 +1,33 @@
+export { getDb, initDatabase, _initTestDatabase } from './init.js';
+export {
+  ChatInfo,
+  storeChatMetadata,
+  updateChatName,
+  getAllChats,
+  getLastGroupSync,
+  setLastGroupSync,
+} from './chats.js';
+export {
+  storeMessage,
+  storeMessageDirect,
+  getNewMessages,
+  getMessagesSince,
+} from './messages.js';
+export {
+  createTask,
+  getTaskById,
+  getTasksForGroup,
+  getAllTasks,
+  updateTask,
+  deleteTask,
+  getDueTasks,
+  updateTaskAfterRun,
+  logTaskRun,
+} from './tasks.js';
+export { getSession, setSession, getAllSessions } from './sessions.js';
+export {
+  getRegisteredGroup,
+  setRegisteredGroup,
+  getAllRegisteredGroups,
+} from './groups.js';
+export { getRouterState, setRouterState } from './router-state.js';
