+++ src/db/router-state.ts
@@ -1,0 +1,16 @@
+import { getDb } from './init.js';
+
+export function getRouterState(key: string): string | undefined {
+  const db = getDb();
+  const row = db
+    .prepare('SELECT value FROM router_state WHERE key = ?')
+    .get(key) as { value: string } | undefined;
+  return row?.value;
+}
+
+export function setRouterState(key: string, value: string): void {
+  const db = getDb();
+  db.prepare(
+    'INSERT OR REPLACE INTO router_state (key, value) VALUES (?, ?)',
+  ).run(key, value);
+}
